refactor: rename pot folder to elements
